Split make_queue models into machine and reservation
