+++ make_queue/models/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ["course", "models"]
+__all__ = ['course', 'machine', 'reservation']
 
 for _import in __all__:
     __import__(f"{__package__}.{_import}")
