+++ make_queue/tests/forms/test_reservation_forms.py
@@ -5,7 +5,7 @@ from django.utils import timezone
 
 from news.models import Event, TimePlace
 from ...forms import ReservationForm
-from ...models.models import Machine, MachineType
+from ...models.machine import Machine, MachineType
 
 
 class ReservationFormTest(TestCase):
