+++ make_queue/tests/models/test_machines.py
@@ -6,7 +6,8 @@ from django.utils import timezone
 
 from users.models import User
 from ...models.course import Printer3DCourse
-from ...models.models import Machine, MachineType, Quota, Reservation
+from ...models.machine import Machine, MachineType
+from ...models.reservation import Quota, Reservation
 
 
 class TestGenericMachine(TestCase):
