+++ make_queue/tests/test_urls.py
@@ -9,7 +9,8 @@ from news.models import Event, TimePlace
 from users.models import User
 from util.test_utils import MOCK_JPG_FILE, assert_requesting_paths_succeeds
 from ..models.course import Printer3DCourse
-from ..models.models import Machine, MachineType, MachineUsageRule, Quota, Reservation, ReservationRule
+from ..models.machine import Machine, MachineType, MachineUsageRule
+from ..models.reservation import Quota, Reservation, ReservationRule
 
 
 class UrlTests(TestCase):
