+++ make_queue/tests/views/test_my_reservations_view.py
@@ -4,7 +4,8 @@ from django.utils import timezone
 from users.models import User
 from ..utility import request_with_user
 from ...models.course import Printer3DCourse
-from ...models.models import Machine, MachineType, Quota, Reservation
+from ...models.machine import Machine, MachineType
+from ...models.reservation import Quota, Reservation
 from ...views.reservation.overview import MyReservationsView
