+++ make_queue/views/reservation/calendar.py
@@ -1,7 +1,8 @@
 from django.views.generic import TemplateView
 
 from util.locale_utils import year_and_week_to_monday
-from ...models.models import Machine, Quota
+from ...models.machine import Machine
+from ...models.reservation import Quota
 
 
 class ReservationCalendarView(TemplateView):
