+++ make_queue/views/reservation/overview.py
@@ -1,7 +1,7 @@
 from django.db.models import Q
 from django.views.generic import ListView
 
-from ...models.models import Reservation
+from ...models.reservation import Reservation
 
 
 class MyReservationsView(ListView):
