+++ make_queue/views/reservation/reservation.py
@@ -13,7 +13,8 @@ from news.models import TimePlace
 from util.locale_utils import timedelta_to_hours
 from util.views import PreventGetRequestsMixin
 from ...forms import FreeSlotForm, ReservationForm
-from ...models.models import Machine, MachineType, Reservation
+from ...models.machine import Machine, MachineType
+from ...models.reservation import Reservation
 from ...templatetags.reservation_extra import calendar_url_reservation, can_delete_reservation, can_mark_reservation_finished
